requirements.txt
absl-py==2.1.0
accelerate==1.1.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==23.2.0
av==13.1.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.3.3
ccimport==0.4.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cmake==3.25.0
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.1
cumm-cu118==0.4.11
cycler==0.12.1
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
easydict==1.13
einops==0.8.0
epic-ops==0.1.0+8af5e3c
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.115.5
fastjsonschema==2.19.1
ffmpy==0.4.0
filelock==3.13.1
fire==0.6.0
flash-attn==2.5.9.post1
Flask==3.0.3
flatbuffers==24.12.23
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.5.0
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.6.0
git-filter-repo==2.45.0
gmssl==3.2.2
google-pasta==0.2.0
gradio==5.6.0
gradio_client==1.4.3
grpcio==1.64.0rc1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.26.2
hydra-core==1.3.2
idna==3.4
importlib_metadata==7.1.0
iopath==0.1.10
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_packaging==0.12.3
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
keras==3.7.0
kiwisolver==1.4.5
kornia==0.7.4
kornia_rs==0.1.4
lark==1.1.9
libclang==18.1.1
lightning==2.2.0
lightning-utilities==0.11.2
lit==15.0.7
lmdb==1.4.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.0.8
msgpack-numpy==0.4.8
multidict==6.1.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.3
nvidia-ml-py==12.535.161
nvitop==1.3.2
omegaconf==2.3.0
open3d @ file:///home/zhangran/desktop/GithubClone/Open3D/build/lib/python_package/pip_package/open3d-0.18.0%2Ba1fb32c92-cp310-cp310-manylinux_2_35_x86_64.whl#sha256=ab8536067342a6a84e4c56695aa596c02a8135042f49c546c08a7ef43a8c0f5e
opencv-python==4.9.0.80
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.11
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pccm==0.4.11
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.2
plotly==5.22.0
pointnet2==0.0.0
pointnet2_ops==3.0.0
portalocker==2.8.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
propcache==0.2.0
protobuf==5.27.0rc3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pybind11==2.12.0
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-multipart==0.0.12
pytorch-lightning==0.7.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
qwen-vl-utils==0.0.8
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.18.1
ruff==0.7.4
safehttpx==0.1.1
safetensors==0.4.3
sapien==2.2.2
schedule==1.2.2
scikit-learn==1.5.0
scipy==1.13.0
semantic-version==2.10.0
Send2Trash==1.8.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
spconv-cu118==2.3.6
stack-data==0.6.3
starlette==0.41.2
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
tensorboard==1.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
timm==1.0.3
tinycss2==1.3.0
tokenizers==0.20.3
tomli==2.0.1
tomlkit==0.12.0
torch==2.0.1+cu118
torch-summary==1.4.5
torch-tb-profiler==0.4.3
torch_scatter==2.1.2
torchaudio==2.0.2+cu118
torchdata==0.7.1
torchinfo==1.8.0
torchmetrics==1.4.0.post0
torchvision==0.15.2+cu118
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.46.2
transforms3d==0.4.2
triton==2.0.0
typer==0.13.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.13
uvicorn==0.32.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
widgetsnbextension==4.0.10
wrapt==1.17.0
yacs==0.1.8
yapf==0.40.2
yarl==1.17.1
zipp==3.18.1
